informer web: refactoring, journal

+++ webapps/informer/java/mobi/eyeline/informer/admin/filesystem/FileSystemSingleHA.java
@@ -14,11 +14,12 @@ import java.io.OutputStream;
 class FileSystemSingleHA extends FileSystem {
 
   @Override
-  public OutputStream getOutputStream(File file) throws AdminException {
-    assert file != null : "Some arguments are null";        // todo почему assert? Это же публичный метод. Тет правильнее IllegalArgument кидать или NullPointer
-                                                            // todo assert надо кидать в приватных методах.  
+  public OutputStream getOutputStream(File file, boolean append) throws AdminException {
+    if(file == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     try {
-      return new FileOutputStream(file);
+      return new FileOutputStream(file, append);
     } catch (IOException e) {
       throw new FileSystemException("io_error",e);
     }
@@ -26,7 +27,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void rename(File file, File toFile) throws AdminException {
-    assert file != null && toFile != null : "Some arguments are null";
+    if(file == null || toFile == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     if (!file.renameTo(toFile)) {
       String err = "Can't rename file '" + file.getAbsolutePath() + "' to '" + toFile.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -35,7 +38,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void copy(File file, File toFile) throws AdminException {
-    assert file != null && toFile != null : "Some arguments are null";
+    if(file == null || toFile == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     try {
       FileUtils.copyFileTo(file, toFile);
     } catch (IOException e) {
@@ -45,7 +50,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void delete(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if(file == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     if (!file.delete() && file.exists()) {
       String err = "Can't remove file '" + file.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -55,7 +62,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void mkdirs(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if(file == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     if (!file.mkdirs() && !file.exists()) {
       String err = "Can't create dirs for '" + file.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -64,7 +73,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public boolean exists(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if(file == null) {
+      throw new IllegalArgumentException("Some arguments are null");
+    }
     return file.exists();
   }
 }
